Simple fix to make sure the skill's mastery is computed correctly.
The skill took all FixtureRecipes as Recipes, so the mastery total wasn't updated correctly (mastery pool was huge, skill had many uncompletable mastery levels in the completion log). This is a fix to overwrite the computeTotalMasteryActions for this skill and check so that only the artisan recipes count as masterable.